+++ src/ResourceManager/Automation/Commands.Automation/Cmdlet/GetAzureAutomationJobOutput.cs
@@ -23,7 +23,7 @@ namespace Microsoft.Azure.Commands.Automation.Cmdlet
     /// <summary>
     /// Gets azure automation variables for a given account.
     /// </summary>
-    [Cmdlet(VerbsCommon.Get, "AzureRmAutomationJobOutput")]
+    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "AutomationJobOutput")]
     [OutputType(typeof(JobStream))]
     public class GetAzureAutomationJobOutput : AzureAutomationBaseCmdlet
     {