+++ src/ResourceManager/DataFactories/Commands.DataFactories/LinkedServices/GetAzureDataFactoryLinkedServiceCommand.cs
@@ -22,7 +22,7 @@ using System.Security.Permissions;
 
 namespace Microsoft.Azure.Commands.DataFactories
 {
-    [Cmdlet(VerbsCommon.Get, Constants.LinkedService, DefaultParameterSetName = ByFactoryName), OutputType(typeof(PSLinkedService))]
+    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "DataFactoryLinkedService", DefaultParameterSetName = ByFactoryName), OutputType(typeof(PSLinkedService))]
     public class GetAzureDataFactoryLinkedServiceCommand : DataFactoryBaseCmdlet
     {
         [Parameter(ParameterSetName = ByFactoryObject, Position = 0, Mandatory = true, ValueFromPipelineByPropertyName = true,