+++ src/ResourceManager/Resources/Commands.Resources.Test/ScenarioTests/SubscriptionLevelDeploymentTests.cs
@@ -29,22 +29,22 @@ namespace Microsoft.Azure.Commands.Resources.Test.ScenarioTests
             XunitTracingInterceptor.AddToContext(_logger);
         }
 
-        [Fact]
-        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Fact(Skip = "Flaky test. Needs investigation.")]
+        [Trait(Category.AcceptanceType, Category.Flaky)]
         public void TestDeploymentEndToEnd()
         {
             ResourcesController.NewInstance.RunPsTest(_logger, "Test-DeploymentEndToEnd");
         }
 
-        [Fact]
-        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Fact(Skip = "Flaky test. Needs investigation.")]
+        [Trait(Category.AcceptanceType, Category.Flaky)]
         public void TestDeploymentAsJob()
         {
             ResourcesController.NewInstance.RunPsTest(_logger, "Test-DeploymentAsJob");
         }
 
-        [Fact]
-        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Fact(Skip = "Flaky test. Needs investigation.")]
+        [Trait(Category.AcceptanceType, Category.Flaky)]
         public void TestStopDeployment()
         {
             ResourcesController.NewInstance.RunPsTest(_logger, "Test-StopDeployment");