+++ src/ResourceManager/DataLakeAnalytics/Commands.DataLakeAnalytics/ChangeLog.md
@@ -19,6 +19,13 @@
 -->
 ## Current Release
 
+## Version 5.1.4
+* Fixed issue with default resource groups not being set.
+* Updated common runtime assemblies
+
+## Version 5.1.3
+* Fixed issue with default resource groups not being set.
+
 ## Version 5.1.2
 * Updated to the latest version of the Azure ClientRuntime.