+++ src/ResourceManager/HDInsight/Commands.HDInsight/ChangeLog.md
@@ -19,6 +19,13 @@
 -->
 ## Current Release
 
+## Version 4.1.8
+* Fixed issue with default resource groups not being set.
+* Updated common runtime assemblies
+
+## Version 4.1.7
+* Fixed issue with default resource groups not being set.
+
 ## Version 4.1.6
 * Updated to the latest version of the Azure ClientRuntime.